fix: filename to kebab-case
