+++ packages/surveys/src/components/general/auto-close-progress-bar.tsx
