+++ packages/surveys/src/components/general/globe-icon.tsx
