+++ packages/surveys/src/components/general/language-switch.tsx
@@ -1,8 +1,8 @@
-import { useRef, useState } from "react";
+import { useRef, useState } from "preact/hooks";
 import { getLanguageLabel } from "@formbricks/lib/i18n/utils";
 import { useClickOutside } from "@formbricks/lib/utils/hooks/useClickOutside";
 import { type TSurveyLanguage } from "@formbricks/types/surveys/types";
-import { GlobeIcon } from "@/components/general/GlobeIcon";
+import { GlobeIcon } from "@/components/general/globe-icon";
 
 interface LanguageSwitchProps {
   surveyLanguages: TSurveyLanguage[];
@@ -49,21 +49,21 @@ export function LanguageSwitch({
         <GlobeIcon className="fb-text-heading fb-h-5 fb-w-5 fb-p-0.5" />
       </button>
       {showLanguageDropdown ? <div
-          className="fb-bg-brand fb-text-on-brand fb-absolute fb-right-8 fb-top-10 fb-space-y-2 fb-rounded-md fb-p-2 fb-text-xs"
-          ref={languageDropdownRef}>
-          {surveyLanguages.map((surveyLanguage) => {
-            if (!surveyLanguage.enabled) return;
-            return (
-              <button
-                key={surveyLanguage.language.id}
-                type="button"
-                className="fb-block fb-w-full fb-p-1.5 fb-text-left hover:fb-opacity-80"
-                onClick={() => { changeLanguage(surveyLanguage.language.code); }}>
-                {getLanguageLabel(surveyLanguage.language.code, "en-US")}
-              </button>
-            );
-          })}
-        </div> : null}
+        className="fb-bg-brand fb-text-on-brand fb-absolute fb-right-8 fb-top-10 fb-space-y-2 fb-rounded-md fb-p-2 fb-text-xs"
+        ref={languageDropdownRef}>
+        {surveyLanguages.map((surveyLanguage) => {
+          if (!surveyLanguage.enabled) return;
+          return (
+            <button
+              key={surveyLanguage.language.id}
+              type="button"
+              className="fb-block fb-w-full fb-p-1.5 fb-text-left hover:fb-opacity-80"
+              onClick={() => { changeLanguage(surveyLanguage.language.code); }}>
+              {getLanguageLabel(surveyLanguage.language.code, "en-US")}
+            </button>
+          );
+        })}
+      </div> : null}
     </div>
   );
 }
