+++ packages/surveys/src/components/general/question-media.tsx
