+++ packages/surveys/src/components/general/survey-close-button.tsx
