+++ packages/surveys/src/components/questions/cal-question.tsx
@@ -1,13 +1,13 @@
 import { useCallback, useState } from "preact/hooks";
 import { getLocalizedValue } from "@formbricks/lib/i18n/utils";
-import { type TResponseData , type TResponseTtc } from "@formbricks/types/responses";
+import { type TResponseData, type TResponseTtc } from "@formbricks/types/responses";
 import { type TSurveyCalQuestion, type TSurveyQuestionId } from "@formbricks/types/surveys/types";
-import { BackButton } from "@/components/buttons/BackButton";
-import { SubmitButton } from "@/components/buttons/SubmitButton";
-import { CalEmbed } from "@/components/general/CalEmbed";
-import { Headline } from "@/components/general/Headline";
-import { QuestionMedia } from "@/components/general/QuestionMedia";
-import { Subheader } from "@/components/general/Subheader";
+import { BackButton } from "@/components/buttons/back-button";
+import { SubmitButton } from "@/components/buttons/submit-button";
+import { CalEmbed } from "@/components/general/cal-embed";
+import { Headline } from "@/components/general/headline";
+import { QuestionMedia } from "@/components/general/question-media";
+import { Subheader } from "@/components/general/subheader";
 import { ScrollableContainer } from "@/components/wrappers/ScrollableContainer";
 import { getUpdatedTtc, useTtc } from "@/lib/ttc";
 
@@ -81,7 +81,7 @@ export function CalQuestion({
             questionId={question.id}
           />
           {errorMessage ? <span className="fb-text-red-500">{errorMessage}</span> : null}
-            <CalEmbed key={question.id} question={question} onSuccessfulBooking={onSuccessfulBooking} />
+          <CalEmbed key={question.id} question={question} onSuccessfulBooking={onSuccessfulBooking} />
         </div>
       </ScrollableContainer>
       <div className="fb-flex fb-flex-row-reverse fb-w-full fb-justify-between fb-px-6 fb-py-4">
