+++ packages/surveys/src/components/questions/multiple-choice-single-question.tsx
@@ -2,11 +2,11 @@ import { useEffect, useMemo, useRef, useState } from "preact/hooks";
 import { getLocalizedValue } from "@formbricks/lib/i18n/utils";
 import { type TResponseData, type TResponseTtc } from "@formbricks/types/responses";
 import type { TSurveyMultipleChoiceQuestion, TSurveyQuestionId } from "@formbricks/types/surveys/types";
-import { BackButton } from "@/components/buttons/BackButton";
-import { SubmitButton } from "@/components/buttons/SubmitButton";
-import { Headline } from "@/components/general/Headline";
-import { QuestionMedia } from "@/components/general/QuestionMedia";
-import { Subheader } from "@/components/general/Subheader";
+import { BackButton } from "@/components/buttons/back-button";
+import { SubmitButton } from "@/components/buttons/submit-button";
+import { Headline } from "@/components/general/headline";
+import { QuestionMedia } from "@/components/general/question-media";
+import { Subheader } from "@/components/general/subheader";
 import { ScrollableContainer } from "@/components/wrappers/ScrollableContainer";
 import { getUpdatedTtc, useTtc } from "@/lib/ttc";
 import { cn, getShuffledChoicesIds } from "@/lib/utils";
@@ -174,62 +174,62 @@ export function MultipleChoiceSingleQuestion({
                   );
                 })}
                 {otherOption ? <label
-                    dir="auto"
-                    tabIndex={isCurrent ? 0 : -1}
-                    className={cn(
-                      value === getLocalizedValue(otherOption.label, languageCode)
-                        ? "fb-border-brand fb-bg-input-bg-selected fb-z-10"
-                        : "fb-border-border",
-                      "fb-text-heading focus-within:fb-border-brand fb-bg-input-bg focus-within:fb-bg-input-bg-selected hover:fb-bg-input-bg-selected fb-rounded-custom fb-relative fb-flex fb-cursor-pointer fb-flex-col fb-border fb-p-4 focus:fb-outline-none"
-                    )}
-                    onKeyDown={(e) => {
-                      // Accessibility: if spacebar was pressed pass this down to the input
-                      if (e.key === " ") {
-                        if (otherSelected) return;
-                        document.getElementById(otherOption.id)?.click();
-                        document.getElementById(otherOption.id)?.focus();
-                      }
-                    }}>
-                    <span className="fb-flex fb-items-center fb-text-sm">
-                      <input
-                        tabIndex={-1}
-                        dir="auto"
-                        type="radio"
-                        id={otherOption.id}
-                        name={question.id}
-                        value={getLocalizedValue(otherOption.label, languageCode)}
-                        className="fb-border-brand fb-text-brand fb-h-4 fb-w-4 fb-border focus:fb-ring-0 focus:fb-ring-offset-0"
-                        aria-labelledby={`${otherOption.id}-label`}
-                        onChange={() => {
-                          setOtherSelected(!otherSelected);
-                          onChange({ [question.id]: "" });
-                        }}
-                        checked={otherSelected}
-                      />
-                      <span
-                        id={`${otherOption.id}-label`}
-                        className="fb-ml-3 fb-mr-3 fb-grow fb-font-medium"
-                        dir="auto">
-                        {getLocalizedValue(otherOption.label, languageCode)}
-                      </span>
+                  dir="auto"
+                  tabIndex={isCurrent ? 0 : -1}
+                  className={cn(
+                    value === getLocalizedValue(otherOption.label, languageCode)
+                      ? "fb-border-brand fb-bg-input-bg-selected fb-z-10"
+                      : "fb-border-border",
+                    "fb-text-heading focus-within:fb-border-brand fb-bg-input-bg focus-within:fb-bg-input-bg-selected hover:fb-bg-input-bg-selected fb-rounded-custom fb-relative fb-flex fb-cursor-pointer fb-flex-col fb-border fb-p-4 focus:fb-outline-none"
+                  )}
+                  onKeyDown={(e) => {
+                    // Accessibility: if spacebar was pressed pass this down to the input
+                    if (e.key === " ") {
+                      if (otherSelected) return;
+                      document.getElementById(otherOption.id)?.click();
+                      document.getElementById(otherOption.id)?.focus();
+                    }
+                  }}>
+                  <span className="fb-flex fb-items-center fb-text-sm">
+                    <input
+                      tabIndex={-1}
+                      dir="auto"
+                      type="radio"
+                      id={otherOption.id}
+                      name={question.id}
+                      value={getLocalizedValue(otherOption.label, languageCode)}
+                      className="fb-border-brand fb-text-brand fb-h-4 fb-w-4 fb-border focus:fb-ring-0 focus:fb-ring-offset-0"
+                      aria-labelledby={`${otherOption.id}-label`}
+                      onChange={() => {
+                        setOtherSelected(!otherSelected);
+                        onChange({ [question.id]: "" });
+                      }}
+                      checked={otherSelected}
+                    />
+                    <span
+                      id={`${otherOption.id}-label`}
+                      className="fb-ml-3 fb-mr-3 fb-grow fb-font-medium"
+                      dir="auto">
+                      {getLocalizedValue(otherOption.label, languageCode)}
                     </span>
-                    {otherSelected ? <input
-                        ref={otherSpecify}
-                        id={`${otherOption.id}-label`}
-                        dir="auto"
-                        name={question.id}
-                        value={value}
-                        onChange={(e) => {
-                          onChange({ [question.id]: e.currentTarget.value });
-                        }}
-                        className="placeholder:fb-text-placeholder fb-border-border fb-bg-survey-bg fb-text-heading focus:fb-ring-focus fb-rounded-custom fb-mt-3 fb-flex fb-h-10 fb-w-full fb-border fb-px-3 fb-py-2 fb-text-sm focus:fb-outline-none focus:fb-ring-2 focus:fb-ring-offset-2 disabled:fb-cursor-not-allowed disabled:fb-opacity-50"
-                        placeholder={
-                          getLocalizedValue(question.otherOptionPlaceholder, languageCode) ?? "Please specify"
-                        }
-                        required={question.required}
-                        aria-labelledby={`${otherOption.id}-label`}
-                      /> : null}
-                  </label> : null}
+                  </span>
+                  {otherSelected ? <input
+                    ref={otherSpecify}
+                    id={`${otherOption.id}-label`}
+                    dir="auto"
+                    name={question.id}
+                    value={value}
+                    onChange={(e) => {
+                      onChange({ [question.id]: e.currentTarget.value });
+                    }}
+                    className="placeholder:fb-text-placeholder fb-border-border fb-bg-survey-bg fb-text-heading focus:fb-ring-focus fb-rounded-custom fb-mt-3 fb-flex fb-h-10 fb-w-full fb-border fb-px-3 fb-py-2 fb-text-sm focus:fb-outline-none focus:fb-ring-2 focus:fb-ring-offset-2 disabled:fb-cursor-not-allowed disabled:fb-opacity-50"
+                    placeholder={
+                      getLocalizedValue(question.otherOptionPlaceholder, languageCode) ?? "Please specify"
+                    }
+                    required={question.required}
+                    aria-labelledby={`${otherOption.id}-label`}
+                  /> : null}
+                </label> : null}
               </div>
             </fieldset>
           </div>
