+++ packages/surveys/src/index.ts
@@ -1,7 +1,7 @@
 import { h, render } from "preact";
 import { type SurveyInlineProps, type SurveyModalProps } from "@formbricks/types/formbricks-surveys";
-import { SurveyInline } from "@/components/general/SurveyInline";
-import { SurveyModal } from "@/components/general/SurveyModal";
+import { SurveyInline } from "@/components/general/survey-inline";
+import { SurveyModal } from "@/components/general/survey-modal";
 import { addCustomThemeToDom, addStylesToDom } from "@/lib/styles";
 
 export const renderSurveyInline = (props: SurveyInlineProps) => {
